setting the reset password endpoint
