+++ .gitignore
@@ -16,3 +16,6 @@ node_modules/
 
 #git-old 
 .git-old/
+
+#DATABASE 
+db/
